Fix 'words'
Originally the 'pattern_to_eval' was set to '<cword>', but currently the whole
group can be updated instead of standalone lighters, thus every lighter got
the current word under the cursor as pattern when '\k' was hit.

TODO:
Adding a new word in another window highlights all of the words in the group.
Solution: Highlight standalone group members.