Rethink group init
If a group is not autoloaded then it has to be in g:high_lighters since
there's no other way to define a group. Thus the conditions of customizing a
non-autoloaded group are dropped.

Group init != group has members (as 'words' group); store the state of init in
a variable.